+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to the gtk-fortran project are documented in this file. The 
 ### Added
 - `src/extract_hl_doc.py`: generates markdown files for the HL gtk-fortran documentation.
 - Intel ifx compiler is now supported by CMake (>=3.20), with the id `IntelLLVM`.
-
+- `tests_gtk_sup.f90`: for testing functions of the gtk_sup module.
 
 ## [gtk-fortran 4.2.1] 2022-04-24